Add ability to send score to database, #27
Add ability to send highscore to a firebase database after finishing
a game. The player can enter a name and send in the score.
Currently the highscores cannot be viewed anywhere. I have implemented
a funciton to retrieve highscores from the database and parse
them into an arraylist of HighscoreObjects.