Fixes issue with highscores overriding, #36
We now generate a random long as an ID according to the structure
described in issue #36. Appearently Unity/C# doesn't have a built in
.NextLong() or equivalent (seriously?) so I made one myself. Not sure
if it's an equal distrubution but it's good enough.

Also made a small change to the submit highscore button. You can now
submit high score by pressing the button again instead of 'enter' in
the input field.