Add lerp in input to make movement smoother, #30
Linear interpolation between the last and current x-input makes the
movement, especially on mobile, smoother but it still looks a bit
strange on mobile when you do really small movements. This happends
mostly when you try to keep the player still.

We could maybe fix this by ignoring input if the differens between
inputs is too small? This solution might be problematic though..