Major changes to UI and look, #45 and #44
Changes the look of the main game by changing the colors of the
background and platforms. Also adds the new button types to the game
over screen.

More or less concludes the 'Improve look of the game'-milestone.
Might be some UI bugs here and there that I didn't catch but most of it
looks good.