Add flipplatforms
This commit adds a new platform type to the game, the flipplatform.
Jumping on it rotates the camera 180 degrees. Since PlatformSpawner was a child of the camera,
it would cause the platforms to be spawned below the player. Fixed this by
having the GameController be the same position as the camera but not a child,
created an empty platformspawner gameobject as a child on gamecontroller.

Rules for flipplatforms:

- It will only flip the first time the player collides with it.
- If spawned, there will always be another one to flip you back no more than 4 platforms after. It could be less though.