Changes to firebase datastructure, #104
Implements changes to the strcuture of the data in the database.
Now we generate a userID when the user first opens the app.
That is then that user's node in the database. Under that node will
then be all of that users submitted highscores.

When collecting the highscore list we then take the largest of those
scores under every id-node into account for the top 10.

That way a user can not have more than one of the spots on the top 10.