Add blinking platforms
This adds blinking platforms to the game. When jumped on, the next three platforms
will blink away and then reappear when jumped on.

There are some known bugs. If the blinkplatforms spawn too close to eachother
they will blink away eachothers next three. For know I'm happy to just make that risk
a low risk, but this needs to be considered.